[BA] Add increase memory suggestion on Overview page
Show info about GC time in Gradle daemon for the last build.
Show a button that opens the memory page in preferences where
user can change the daemon memory settings.
If GC percentage was high enough (10% seems reasonable based on
the stats) show it as a warning (add a warn sign).

Bug: 155317033
Test: added
Change-Id: I4c4c17000ae182c5fd6835d7d7c1f6ec513a6e68