Keep render error result when render task creation failed
Currently, if the render task creation fails, a render result is
returned that includes the logger containing the error information
from the failed task creation. But this information is never displayed
as the error panel update is asynchronous and before it runs, the
render result is replaced by a generic empty error render result.
This makes sure that the original error render result is kept, so that
the errors can be displayed in the error panel.

Bug: N/A
Test: Manually checked
Change-Id: Id67620bb496c2893837699dba4d0c42a77468d9b