[workspace model] IDEA-246697 Fix OOM at external build system project import
The precondition, the project should be opened based on this storage already be filled with project data.
The first problem that during reimport we create a lot of modifiable models with copied storage instead
of reusing one storage build in all modifiable models. To fix it `IdeModifiableModelsProviderImpI` was
rewritten and now, if the new project model enabled it creates modifiable models with a single instance
of the storage builder passed as a parameter.
The second problem of the huge memory consumption was related to the `VersionedEntityStorageOnBuilder`
which creates and caches a copy of the store if any changes were there. Thus we get `ValuesCache` the
full of storages of the different versions (e.g during regular execution of
`ModifiableRootModelBridgeImpl.addModuleOrderEntry` on the same modifiable model). To fix it
`DummyVersionedEntityStorage` was introduced which cache nothing and now modifiable model contains a
boolean flag which tells us if the data in the model should be cached.

GitOrigin-RevId: e54ceec0ec4f3b3fb18b68137580dbd4b070e744