Make FakeTraceProcessorService serve real data from TPD
For most of out test we cannot use the real TPD (because of extra time
and because how our code is organized in profilers and
profilersAndroid).

This CL make FakeTraceProcessorService able to read serialized versions
of the content produced by the real TraceProcessorService, this way we
can keep referencing the same traces in our testData and be able to
assert/test based on the actual data that would be produced by a real
TPD instance running.

Also adds SerializedTraceProcessorModelForTestsVerifier that verify
that our current serialized model matches one produced by a real TPD
run.

Bug: 162495274
Test: existing + FakeTraceProcessorServiceTest + SerializedTraceProcessorModelForTestsVerifier
Change-Id: Ifbf01f18a47cf90296674c906352392eb0153edc