IDEA-246929: AndroidBuildProcessParametersProvider should not throw NPE if some class is available via "jrt://..." url
i.e. when some class is provided by the JVM. Note that in some situations ignoring missing JARs may lead to undesired consequences. E.g. if IDEA runs on Java 11 EA version (and javaee-related classes are available via  jrt://), but JPS process runs on Java 11 stable version (and javaee-related classes should be added from some external jar). Currently we consider this situation to be rare. NPE is bad thing anyway. Replace it with a warning in prod (and throw an error in unit test mode), since we cannot do anything helpful anyway. If jar is missing, but needed then ClassNotFound will be thrown, if it is not even needed (e.g. project does not have android modules) error will not happen at all.

GitOrigin-RevId: 2616c9c898d22d397e97082b269f2a6b39b437af