[Dsl] maybe fix test issue on Windows
Three tests are failing on Windows: all three of them (and only those)
construct a new submodule in addition to the two modules constructed
in GradleFileModelTestCase.

A change in February 2019 converted most of the file handling in
GradleFileModelTestCase from IOFiles to VirtualFiles, but left the
construction of the new submodule alone.  The test cases now (August
2020) fail with an error in tearDown, failing to delete a file.  The
intent in this change is to mirror the file handling of the rest of
the test system, in the hope that this addresses the issue.

Bug: none filed as far as I can see
Test: existing
Change-Id: I22f848a1be35b7399b33a4bad04b0e4e5776cf0b