App Inspection: Keep dead processes in the pulldown list
When a process is stopped, users may still want to inspect it.

With this CL, a process is kept in the pulldown list when it is stopped,
but as soon as it is replaced with a new process (e.g. a new preferred
process is started up), it will be removed.

This behavior mimics that in logcat.

Fixes: 154951817
Test: Added
Change-Id: Idf85e0cf1ed7807f44da4bcf3b07fd462111aa59