[BA] Add empty selection state to warnings page
Introduce empty details page selection state and stop
forcing selection of the detail page in warnings.
Empty detail page shows a message to select page in a tree.
It is required for next CL that introduces filters to
warnings page making it possible for currently selected
page disappear. It is better to default to empty selection
in this case than trying to figure out what node to select
instead.

Bug: 150297387
Bug: 150295612
Test: Updated
Change-Id: I52afbcf3d417a155b1e3a9e09316926dc24b00b0