IDEA-247946: remove dependency `intellij.android.buildCommon` > `intellij.android.common`
`android.buildCommon` is used in the JPS build process. `android.common` has a lot of dependencies which are not available in JPS (available in the IDE only). In order to avoid potential NoClassDefFoundError, just replace `BufferingFileWrapper` with `FileWrapper` and remove the dependency. There is no performance penalty for this change, because `ProjectProperties.parsePropertyFile` will wrap the `InputStream` with the `BufferedReader`, so we just removed double-buffering (which is good in 3D games, but often useless when reading files)

GitOrigin-RevId: 3367e4ff689e5422f37fcacf0f22e5dddfd03983