159532832: Android Studio 4.2's lint error UI is confusing
In Studio 4.2, all errors identified by lint were showing
up with the "unknown symbol" style (text color red, instead
of a red underline). This seems to be due to a new
mapping from highlight styles to problem styles in the
platform code.

To fix this, we map directly to the preferred error style.
While we're at it, we might as well make the unknown id
style continue to use the unknown font style.

Test: Existing
Fixes: 159532832
Change-Id: I359a01fb4f225d5890e26341add0ecc88d02f61e