Do not access IDevice on the EDT
GradleModuleSystem.getApkProvider() do need a device spec here and
it make take some time to construct such from an IDevice instance.

Instead GradleModuleSystem can get the device version when processing
`getApks()` request which runs in the background.

This change applies only minimal changes required to NOT construct
an instance of AndroidDeviceSpec. It passes `null` value instead as it
is unused now.

Bug: 163631578
Test: n/a
Change-Id: Ic1e28990d486efb751f109ecf47eea35768bc584