Add support to stamp the build timestamp
The way of doing this in bazel is to depend on the volatile
status file. This file contains the timestamp, and it does
not trigger a rebuild when it changes (which it does on
every build). This way the tool is determinitic (can be tested)
and the same timestamp can be used across different files.

Test: added
Bug: 156497102
Change-Id: I03548f653f908fde682a8c3947c2427b3d7df387