Inline AndroidModule.moduleDependencies property
which is used to pass values between two instance instance methods which
are called one after the other.

Note, that currently `moduleDependencies` is always empty before the
call to `addSelectedVariant` because we skip any previously visited
modules (even if it could happen that two different modules depend on
different variants of the other module we do not support it in sync).

Next steps:
- eliminating recursion in addSelectedVariant
- simplifying AndroidModule and removing "unused" properties
- reworking AndroidModule to hold Ide models

Bug: n/a
Test: n/a (existing sync tests and synced project snapshots)
Change-Id: I644b428d5d981b7c504b099d4404b91c31c39c0c