[Dsl] make modulePaths usably efficient
Firstly, modulePaths is often used for membership rather than
ordering, which makes a List implementation suboptimal:
innocuous-looking operations are necessarily O(N) from use of
List.contains().

However, making modulePaths() return a Set or some other O(1ish)
membership container does not help if constructing that set takes O(N)
for each call, as then something such as constructing a module
tree (involving computing the parent for each module) will take
O(N-squared) time from the N calls.

Amortize the cost of construction by cacheing the result, based on the
Dsl's modification and committed counters.  The cache is not
updated (though the value is recomputed) if the counters differ,
corresponding to an intermediate state during modification; when not
in such an intermediate state, operations such as parentModule() and
buildFile() should now be amortized-O(1) in the number of modules.

Fixes: 162259883
Bug: 149207197
Test: existing correctness tests
Test: manual check that the cache is hit in large projects
Change-Id: I56aaa004976eaad822b6d12d005785028b66b880