Don't wrap text in Animation Inspector properties panel
We were wrapping the text in the animation properties panel to make sure it
was always displayed to the users. A side effect, however, is having the
text "jumping", which might make it hard for users to track a particular
value. In this CL, we removed the text wrapping in the panel and moved the
property value so it's displayed below the label, not next to it. This way,
users can see at least the beginning of the property label and value, and
can use a scrollbar when there is a very long label or value.

Fixes: 163877630
Test: N/A. Visual Change.
Change-Id: I44989b1fa0996f247c64f8fa058b0a6f32fb131a