Add query history button to SQLite evaluator
Adds a button that when clicked shows a list of queries
previously executed by the user.

Each time a list item is selected in the list's selection
model, it's value is shown temporarily in the
editor text field, highlighted in blue.
Only when the user presses enter the value is added
permanently to the editor.

Test: SqliteEvaluatorControllerTest.kt, QueryHistoryViewTest.kt
Bug: 167369710
Change-Id: I6c4a0ae0142c6d5bd79296132840dfd69c8fea44