Fix slowness of AndroidResourceFilesListener.getFilesToProcess
The AndroidResourceFilesListener.getFilesToProcess may be called on the UI thread and has to be fast. In order to achieve that, we use an approximate check of file relevance that may produce false positives but is guaranteed to not perform any file I/O.

Also streamlined code inside the AndroidFileChangeListener.isRelevantFile(VirtualFile) method.

Test: existing
Bug: 166646527
Change-Id: I3b4d12d4cbeea4241e845d80e4ca60263d96bee4