Use annotation index instead of visitors
This allows the annotation finding to be much faster and stop relying on
the slow UAST visiting. It also allows to stop relying on the hack of
looking at the imports to detect if the Preview panel should be present
by also using the index.

This removes the @Slow annotation from the annotation finder and also
simplify the method by moving to a non-uast interface.

Fixes: 166644307
Test: Existing tests pass
Change-Id: I9baf640032875f8cf71d2a37e74f361f7b15c74b