Enable debugger selection for empty run configs
Currently when a user clicks on "Attach Debugger to Android Process"
with no run configurations defined in the project the debugger selection
combobox will be disabled even though it's supposed to be enabled when
there are no run configurations. The user has to manually re-select the
"[Create new]" option to enable the debugger selection combobox.

This CL ensures the debugger selection combobox is enabled correctly
when process chooser dialog opens with no run configs.

Bug: N/A
Test: Opening the process chooser enables debugger selection
correctly.

Change-Id: I6198a26345a66f1911e5e6c6a0d94d9a7b64bbf5