[openapi] IDEA-248714 - html link in Project Structure -> Problems
Both ProjectStructureProblemDescription#getDescription and ConfigurationError#getDescription used to return a string with caused problems with escaped html characters that can be converted multiple times and hence appear not as html tags on panes but as strings with extra symbols. In order to overcome these problems this patch changed the return type of the methods to HtmlChunk to provide the type safety when working with html tags.

(cherry picked from commit 480ecf2ca0b8dcc2112139fa84d3e788ed158cd0)

Change-Id: Ie208c09669b1345f6fbf79a0d44fafda48821e4e