Disable flag for "Work Manager Inspector"
This flag was enabled for canary only. For beta and stable, we want
to revert the change.

Unfortunately, it's not a simple flag flip, because we also have to
tweak the behavior of the app inspection tool window when we revert
back to one inspector instead of two, e.g. rename "app inspection"
back to "database inspection" and revert icons.

Bug: N/A
Change-Id: I8d3a3d0b6ea3d5c4253edbb48a3d3885e79fe2eb
Test: N/A