Wait 1 second after sync for async events
... on testExceptionsCreateFailedBuildFinishedEvent.

The test had a time out and waited for at least 3 events (one for start,
one for progress and one when finished). But the test should take all
the events, not only the first 3.

This change makes the test to wait one second and then verifies that at
least 2 events were created before processing them.

Bug: 161618318
Test: This change enables a test.
Change-Id: I54142b3b6561a9279f07cd8aedee5a90d37bb3c6