Extend EmptyStatePanel to support actions
Sometimes when showing the empty state panel, instead of
having a link to a web page it is useful to show a link
that when click runs some code in studio.

This change extends EmptyStatePanel to support that use case
by changing UrlInstruction to HyperlinkInstructions, that adds
the possibility to add an HyperlinkListener to the label.

Test: InstructionsPanelTest.kt
Bug: 168793024
Change-Id: I65116d5c77c98181636319c7bd647f799d61d97c