Handle end of stream case properly
Currently, if the TPD process output stream end before we detect one of
the two matches we look for, the code will be stuck forever in the
STARTING status.

This CL reworks the mechanism a bit and add a new END_OF_STREAM status
that for TPD spawning is treated as the spawn process failed (for
example, the process died and isn't producing anything else).

Bug: 169033108
Test: TraceProcessorDaemonManagerTest
Change-Id: I520c3da657903b978dedefd53e4ec1ac4598e846