Refactor AppInspectionView and its test code
This changes the way tabs are added to the tool window. With this CL, tabs will be added immediately to the tool window with a loading message. Once the inspector tab is ready to be rendered, then it will be added into the tab. Non-applicable tabs are not displayed (just like before).

This change also converts the existing testing hooks in AppInspectionView to coroutine, which makes testing easier.

Test: updated existing tests.
Bug: 165319681
Change-Id: I5c64fc51a589c2b424ff4dc55dca53f5388b4015