Fix multiple target device build optimizations.
Don't combine device sdk versions if they differ when
deploying to multiple devices.

Optimizationions to target a device of a
particular level might not be safe for devices above that level,
for example see https://issuetracker.google.com/163791525.

If we want to re-introduce this, perhaps we should have a new field that
includes all the API levels of the devices togher.

Fixes: 163791525
Test: Updated unit tests AndroidDeviceSpecUtilTest and MakeBeforeRunTaskProviderTest
Change-Id: I7d317869b8bfdb6fced2a95987aca3c674d049c1