Use GetLibraryVersionsCommand when launching inspectors
AppInspectionView will now start using GetLibraryVersionsCommand to
check the compatibility of all potential inspector tabs before launching
them. This means some inspectors need not to be launched, and we can
show an info tab right away. This is in contrast to previously where we
waited until after createInspector command came back with an error.

Test: AppInspectorLaunchSupportTest
Bug: 165319681
Change-Id: I8e8b9d47d9be22a34853971833e3d0cd746cc8ea