Set debugger selection based on debugger ID.
AndroidProcessChooserDialog sets the selected type of debugger
automatically using the debugger returned from
AndroidDebuggerInfoProvider#getSelectedAndroidDebugger which isn't
guaranteed to return the same object instance as one of the debuggers
from AndroidDebuggerInfoProvider#getAndroidDebuggers.
JComboBox#setSelectedItem will only set the selected item if it's the
same object.  This means the process chooser dialog may not set the
correct debugger based on luck of implementation.

This CL changes the AndroidProcessChooserDialog to set the selected
debugger based on debugger ID instead. This respects the API of
AndroidDebuggerInfoProvider and always works.

Bug: 169976577
Test: Manually verified. I could extract it out and write a test for it
but it will just be repeating the implementation.

Change-Id: I1a5232f4220d1d4dd44bf51d69a49b943ceaf147