Add a JUnit run configuration template
This change takes advantage of the the ability to share template run configurations introduced in IJ 2020.3 to automate few configuration steps when switching to the unbundled project.

User has to define the TEST_TEMP path variable pointing to /tmp/idea-test on Linux and Mac, or to %TEMP%\idea-test on Windows, where %TEMP% is the value of the TEMP environment variable.

Test: N/A
Bug: N/A
Change-Id: Ic7cc351b6cbae001ed071d625524389060b3ae2d