Add message panel to new query view
In the new query view we sometimes display a message
to the user. So far we've used the empty text of the
table. This CL improves this solution so that we can
show a panel specifc for showing messages, instead of
using the empty text of the table.

Test: SqliteEvaluatorControllerTest.kt, SqliteEvaluatorViewImplTest.kt
Bug: 169387231
Change-Id: Ibe7fee2f44936b2e84901c15b461be238dd9f503
(cherry picked from commit 071fbb80aadccdba7eb27f8473d34feb5aa9283c)