Do not clear ResourceClassRegistry cache when resetting ids
Resetting dynamic ids is performed when doing general cache cleaning,
and in that case it is appropriate to also clear the ResourceClassRegistry
cache. But it can also be performed during the creation of RenderTask, and
there it does not make sense to reset ResourceClassRegistry.

Bug: 154470366
Test: checked manually
Change-Id: Id1983435e48ff4019ea791933ae80dd99671e09f
(cherry picked from commit b389fa6ddda47d1047d44e8f97fd5bba193572f6)