[lint] restore signature of checkDslPropertyAssignment
It is sometimes valid to assign to global, effectively (syntactically)
parentless properties in Gradle build scripts.  A natural
representation for this would be to consider the parent to be null,
and a previous change in the 4.2 cycle made a change to
checkDslPropertyAssignment to allow this.  However, this broke
compatibility with first- or third-party lint Gradle build script
checkers, such as the one for ActivityResultFragmentVersionDetector
from the Android platform, or JcenterDetector from vanniktech on
github.

Adapt the callers of checkDslPropertyAssignment to perform the null to
empty string conversion where appropriate.

Include a test case which fails if the implementation here is not
changed in sync with the lint API.  The test case added when making
the original change to checkDslPropertyAssignment passes because the
exception thrown from passing a null argument in a non-nullable
position to the method obscures the failure to detect the valid
language level directives.

Bug: 170026127
Test: added
Change-Id: Ib172983c043f2581a53f2a4e4d54c8d3a00bc667