Move reusable functionalities to base stage
This change cut and paste functionality in `MainMemoryProfilerStage`
to its base class so it can be re-used by the allocation stage in a
future change.
It also weakens the dependencies of some components
(e.g. `DetailedMemoryUsage`, `MemoryUsageTooltip`, etc.) that only
depend on those functionalities.

This change also removes uses of `isUnifiedPipelineEnabled`
and assumes it's always `true`.

The change does not introduce any behavior change.

Bug: 169256185
Test: existing

Change-Id: I7dd0005b5236774331885fc11d9ee97037b9d42d