Add delayed surface deactivation
When using the Compose preview, if the preview is not visible we will
start the deactivation process.
The surface will be kept active for 5 minutes in case the user wants to
come back to it. If the surface is deactivated and the user comes backs,
a refresh will be done.

If the user opens more than 5 preview tabs, an LRU queue is used and the
least recently used will be deactivated.

Bug: 169940598
Test: New deactivation queue tests
Change-Id: I1d7323d809676bc7f9c3e9f722364aa0eaa01714