[AUA] convert UI strings to bundle messages
This is almost a pure move change; no change in functionality is
expected.  One small exception is the change to the `target` displayed
text if there is an AGP dependency in a Gradle build file in buildSrc,
but none in any other build file.  (This would be a very strange
situation in the wild.)

Fixes: 161791543
Test: manual
Change-Id: Id1374f33520d53d9ff9af99a254ba577c70013fe