Add PropKeysLabelInspection
Add a lint inspection to show a weak warning when defining Compose animation
PropKeys without setting the `label` parameter, which cause them not to be
properly identified in the Animation Preview, since a generic name will be
displayed in this case.

Add the inspection under a flag, so we can disable it anytime if needed.

Bug: 170314149
Test: AnimationsInspectionsTest

Change-Id: Icaa6e40a3f9e1af91f8ff8c55dbcda82dddb0429