IDEA-252775 [merge-android-41] copy gradle-dsl to community modules after merge
... and restore Android's gradle-dsl sources to their original state (as they are in the AOSP repository for android-plugin 4.1)

Intention is that gradle-dsl will be removed from IJ for now. Only one copy will remain - in android sources. All the dependencies from the gradle-dsl on the Android plugin will be removed, and gradle-dsl will be used by IJ. In far-far future gradle-dsl might be moved to IJ sources again.

GitOrigin-RevId: 87f11a5bd73dfab441a29945cd0b87378801abfb