Quickfix to use AGP 3.1.0+, GRADLE_MINIMUM_VERSION
... when there are old Gradle issues.

We currently detect problems when an old version of Studio is used and
have a couple of quickfixes for those. They only change the version of
Studio but let the AGP version untouched, causing a circular change, for
example, going from Gradle 2.2 to 2.6, then 4.8.1 and then back to 2.2.

This change makes the direct jump to the minimum supported version:
Gradle GRADLE_MINIMUM_VERSION and AGP 3.1.0 (3.2 if using JDK 11). The
new quickfixes do not use the upgrade assistant since it is very likelly
that no successful sync is possible and hence the used AGP version
cannot be get from the model (AUA needs this). A separate change will
try to use it.

Bug: 171583343
Test: New tests in OldAndroidPluginIssueCheckerTest and UpgradeGradleVersionsQuickFixTest.kt
Change-Id: I1747d8b99df422ea03cbfdcbff4e7e1a9e4a4259