[workspace model] IDEA-254450 Speed up Gradle project sync at RootIndex initialization
The root cause of this problem was in constant initialization of RootIndex (`RootIndex#initLibraryClassSourceRoots`)
which was canceled by started `WriteAction`.
The Main scenario which leads to the problem:
During the Gradle project synchronization `ExternalSystemModulePropertyManager#setExternalOptions` code executed and in
New Project Model it requires WriteAction. At the same time, `RootIndex` tried to be initialized under the ReadAction
(`ProgressIndicatorUtils#runActionAndCancelBeforeWrite`) and canceled due to started WriteAction. Due to the execution of
`ExternalSystemModulePropertyManager#setExternalOptions` with each module during the project synchronization, initialization of `RootIndex`
constantly failed. Incomplete initialization leads to WriteAction waiting and thus slow down project synchronization because of the long,
constant ReadAction.
This fix tries to avoid WriteAction at `ExternalSystemModulePropertyManager#setExternalOptions` by reusing existing
`IdeModifiableModelsProviderImpl#getActualStorageBuilder` instead of writing to the main store directly.
Why it was not produced with the old project model?
The old implementation didn't require WriteAction at `ExternalSystemModulePropertyManager#setExternalOptions` and theoretically,
it could lead to some sort of inconsistency.

GitOrigin-RevId: f69442f19109bccaffa82aeaf81dfbd8d1aecdad