Introduce Target
Virtual devices used to contain a list of snapshots. I changed the list
to a single snapshot in Commit a852610aa07da349881fbb9b737627080be5ee4e.
That was a mistake. It's odd that Virtual Device–Snapshot 1 is
considered a different device than Virtual Device–Snapshot 2 and there's
hackery at places like PopupActionGroup.java:87 to compensate for that.

Introduce a new Target abstraction that replaces those device–snapshot
pairs and eventually go back to virtual devices containing lists of
snapshots

This commit is Step 1 in that process

Bug: 120497509
Test: Existing tests
Change-Id: I139cffa536de260768e22b7a2ed363b33f5b0474