[auto-import] Fetch indices from network and generate local registry.
This is just the starting cl for the whole following implementations.
TODO items will be taken care of in the subsequent cls.

In this CL, the original `MavenClassRegistry` is renamed to
`MavenClassRegistryLocal`. Also we introduce an interface here, named
as `MavenClassRegistry`. `MavenClassRegistryRemote` is for fetching
data from `GMavenIndexRepository`, and generating lookup to provide
subsequent quick fixes to those unresovled references.

Bug: 173547479
Test: MavenClassRegistryRemoteTest
Change-Id: I6b36c5b3271fb9fe552012d3333110ad7cf3b6cb