[AUA] capture common necessity patterns
The two standard necessities we have seen so far (other than the
trivial "always applicable") are: a point change, where there is a
discontinuous change in tool behaviour at a given version, and the
necessity for that transform is a function of the current and new AGP
versions relative to that particular version; and a region change,
where a feature's replacement is introduced (and the existing feature
deprecated) at one version, and the deprecated feature is removed at a
later version.  Define and use helper functions capturing those two
patterns.

Bug: n/a
Test: existing (no change in functionality)
Change-Id: Ib390e83a6fb4397326fc1cc17495dc5e1612966c