Fix tooltip in name column of property table
A property table tooltip is found from the deepest component of a
cell renderer under the mouse. For a name column we failed to
implement getToolTip for the labels held by DefaultNameComponent.
This caused the tooptip to only show up in between attr names.
Here is a fix.

Fixes: 174058488
Test: Unit test updated to use the deepest component under mouse
Change-Id: I9c85ea2e38521bcd0f8c5b4ccd3c0a7f63a0fa43
(cherry picked from commit 130cefb25d7818077ba341dc8222230dc62b4643)