Fix testAnimationInspector
Update the test to reflect the (not so) new UX, in which you close the
animation preview by clicking on "Stop Animation Preview", not by toggling
the animation preview toolbar button.

It will remain ignored because we use an older version of Compose in the
tests, and while we don't upgrade it the animations won't be detected.

Bug: 172894609
Test: ComposePreviewTest#testAnimationInspector
Change-Id: I0daba642502b5d2e74656a7372f1a0645d4ca546